+++ Trabalho/MyZoo/src/MedicoVet.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class MedicoVet extends Funcionario {
 
-	public MedicoVet(String nome,int cpf, String email, Date nascimento,
+	public MedicoVet(String nome,long cpf, String email, Date nascimento,
 			double salario, int carga_horaria, long telefone, Time horaDeChegar,
 			Time horaDeSair) {
 		super(nome,cpf, email, nascimento, salario, carga_horaria, telefone,
