+++ Trabalho/MyZoo/src/Pessoa.java
@@ -2,11 +2,11 @@ import java.util.Date;
 public abstract class Pessoa {
 	//Atributos
 	String nome;
-	int cpf;
+	long cpf;
 	String email;
 	Date nascimento;
 	//Construtor
-	public Pessoa(String nome, int cpf, String email, Date nascimento) {
+	public Pessoa(String nome, long cpf, String email, Date nascimento) {
 		super();
 		this.nome = nome;
 		this.cpf = cpf;
